Updated eslint error
